CP-4042: Selectively pass GenID to xenstore
Only Win8/Win12 guests should have GenIDs. If the generation_id is "", we
ignore it. If it is "0:0", we refresh it. If it is not "", then we will pass it
to xenstore. When we need to refresh a GenID (on import, clone, etc.), we only
do so if it is not "".

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>