Handle invalid edition on xapi startup
We changed the editions in a later version of XenServer. If an old xapi
restarts during an upgrade, we may have an invalid edition in the database. If
the edition found in the database isn't in the list of allowed editions, we
instead use the "default edition" (the first edition returned by
V6client.get_editions).

CA-121350

Signed-off-by: Mike McClurg <mike.mcclurg@citrix.com>